backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
networkx==3.2.1
langchain==0.0.329
openai==1.3.3
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.1.0
pydantic-settings==2.0.3
alembic==1.12.1
